Version 29
This version has the updates:
1.	Only support 2 urls settings from the app:
a.	"jctbackend.herokuapp.com"
b.	"cabinet-staging-a663e71fb1a6.herokuapp.com"
Anything else will be fail case and can’t sending any updates information.

2.	Battery status sending using sync request structure
a.	The device has the charger plugged in, you only be able to use “vbatScaled” 1.0 for the plug in status.  To ensure that the charger is plugged into the device, you can use the voltages[15] adc counts for the confirmation.
This is overkill but just in case you want to double check the status of the charger plug in.  When the system has the charger plugs in, the ADC count always higher than 500.
The device has very simple charging circuit so the precise level of the battery during charging can’t reliably report so you have to just threw the charging symbol on the app.
"engrData": {
  "timestamp": "1700597744",
  "voltages": [1438, 1538, 1334, 1229, 1163, 1383, 1473, 1253, 1340, 1131, 1582, 889, 1568, 739, 2796, 630],
  "vbatScaled": 1.0,
  "vbatMeas": 100.0
},
b.	The device has the charger unplugged the value of the “vbatScaled” will be 0.0.  Currently, the “vbatMeas” is reliable report the true current battery level.
When in doubt you can use this formular to calculate the percentage of the battery by using battery ADC counts from voltages[14].  Again, just use the “vbatMeas” value is good enough.
Below is how I calculating the battery level for BlueTooth information.  It is here just for your information only.  No need for you to implement on the backend.

#define BATTERY_OFFSET      (200) //lost 200mV on Q4
#define VOLTAGE_RATIO       (1.4)
#define BATTERY_MAX_ADC     (3700)
#define BATTERY_MAX_CAL     (3700.0)
    //update battery levels
    //Vbatt = (Vmux * 1.4 + 200) mV, assume 3700 is the max
    uint16_t Vmux = voltages[BATTERY_MUX_CHANNEL];
    uint16_t Vbatt = Vmux * VOLTAGE_RATIO + BATTERY_OFFSET;

    if(Vbatt > BATTERY_MAX_ADC) //cap at 3700
    {
        Vbatt = BATTERY_MAX_ADC;
    }

    //battery percentage from 0mV to 3700mV 0% to 100%
    stBatteryLevel.battery_level = (uint8_t)( ((float)Vbatt * 1.0 / BATTERY_MAX_CAL) * 100.0);

"engrData": {
   "timestamp": "1700597788",
   "voltages": [1138, 1533, 1140, 1099, 889, 1156, 1178, 856, 1005, 673, 1017, 471, 775, 353, 2656, 264],
   "vbatScaled": 0.0,
   "vbatMeas": 100.0
},